+++ graph/graph-adapter-input-rest-spring-mvc/src/test/kotlin/eu/tib/orkg/prototype/statements/application/StatementControllerUnitTest.kt
@@ -9,14 +9,14 @@ import eu.tib.orkg.prototype.createLiteral
 import eu.tib.orkg.prototype.createPredicate
 import eu.tib.orkg.prototype.createResource
 import eu.tib.orkg.prototype.createStatement
+import eu.tib.orkg.prototype.pageOf
 import eu.tib.orkg.prototype.spring.spi.FeatureFlagService
-import eu.tib.orkg.prototype.statements.api.BundleConfiguration
 import eu.tib.orkg.prototype.statements.api.StatementUseCases
-import eu.tib.orkg.prototype.statements.domain.model.GeneralStatement
 import eu.tib.orkg.prototype.statements.domain.model.ThingId
 import eu.tib.orkg.prototype.statements.spi.TemplateRepository
 import eu.tib.orkg.prototype.testing.spring.restdocs.RestDocsTest
 import eu.tib.orkg.prototype.testing.annotations.UsesMocking
+import eu.tib.orkg.prototype.testing.spring.restdocs.documentedDeleteRequestTo
 import io.mockk.Runs
 import io.mockk.clearAllMocks
 import io.mockk.confirmVerified
@@ -24,17 +24,16 @@ import io.mockk.every
 import io.mockk.just
 import io.mockk.verify
 import io.mockk.verifyAll
-import java.net.URLEncoder
 import java.security.Principal
 import java.util.*
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.TestInstance.*
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
 import org.springframework.http.MediaType
+import org.springframework.restdocs.request.RequestDocumentation.*
 import org.springframework.security.test.context.support.WithMockUser
 import org.springframework.test.context.ContextConfiguration
 import org.springframework.test.web.servlet.MockMvc
@@ -184,7 +183,8 @@ internal class StatementControllerUnitTest : RestDocsTest("statements") {
     @DisplayName("Given a user is logged in")
     inner class UserLoggedIn {
         @Test
-        fun `when deleting a statement with a literal object and service reports success, then status is 204 NO CONTENT`() {
+        @DisplayName("when deleting a statement with a literal object and service reports success, then status is 204 NO CONTENT")
+        fun deleteLiteralStatement_isNoContent() {
             val s = createResource().copy(label = "one")
             val p = createPredicate().copy(label = "has symbol")
             val l = createLiteral().copy(label = "1")
@@ -194,8 +194,16 @@ internal class StatementControllerUnitTest : RestDocsTest("statements") {
             every { principal.name } returns userId.toString()
             every { statementService.delete(st.id!!) } just Runs
 
-            mockMvc.delete("/api/statements/${st.id}", principal)
+            mockMvc.perform(documentedDeleteRequestTo("/api/statements/{id}", st.id!!).principal(principal))
                 .andExpect(status().isNoContent)
+                .andDo(
+                    documentationHandler.document(
+                        pathParameters(
+                            parameterWithName("id").description("The identifier of the statement to delete.")
+                        )
+                    )
+                )
+                .andDo(generateDefaultDocSnippets())
 
             verify(exactly = 1) { statementService.delete(st.id!!) }
             verifyAll { principal.name }