+++ rest-api-server/src/integrationTest/kotlin/eu/tib/orkg/prototype/statements/application/StatementControllerIntegrationTest.kt
@@ -84,7 +84,7 @@ class StatementControllerIntegrationTest : RestDocumentationBaseTest() {
                 document(
                     snippet,
                     requestParameters(
-                        parameterWithName("page").description("Page number of items to fetch (default: 1)").optional(),
+                        parameterWithName("page").description("Page number of items to fetch (default: 0)").optional(),
                         parameterWithName("items").description("Number of items to fetch per page (default: 10)")
                             .optional(),
                         parameterWithName("sortBy").description("Key to sort by (default: not provided)").optional(),